Fixes for poor gcc
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6835 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/image.cc
@@ -1340,7 +1340,7 @@ ImageA ImageRGBA::Lightness() const {
         if (v <= 0.04045f) {
           table[i] = v / 12.92f;
         } else {
-          table[i] = std::powf((v + 0.055f) / 1.055f, 2.4f);
+          table[i] = std::pow((v + 0.055f) / 1.055f, 2.4f);
         }
       }
       return table;
